fix utility creazione database

+++ Progetto-I/src/UIcapo/buttonLayout.java
@@ -5,6 +5,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -26,6 +27,7 @@ public class buttonLayout extends JPanel{
     String tablename;
     TableModelBuilder tablePane;
     DatabaseManager mgr;
+    JFrame frame = new JFrame();
     
     public buttonLayout(DatabaseManager mgr,HashMap<String,JTextField> map,String tablename,TableModelBuilder tablePane){
         this.tablePane=tablePane;
@@ -67,27 +69,28 @@ public class buttonLayout extends JPanel{
                     if(!areTextFieldEmpty()){
                         PreparedStatement con;
                         try {
-                            if(tablename.equals("impiegati")){
-                                con = mgr.getConnection().prepareStatement("USE ristorante");
-                                con.executeQuery();
-                                con = mgr.getConnection().prepareStatement("INSERT INTO impiegati(username,password,ruolo) VALUES (?,?,?)");
-                                con.setString(1, map.get("username").getText());
-                                con.setString(2, map.get("password").getText());
-                                con.setString(3, map.get("ruolo").getText());
-                                con.executeUpdate();
-                            }else if(tablename.equals("menu")){
-                                con = mgr.getConnection().prepareStatement("USE ristorante");
-                                con.executeQuery();
-                                con = mgr.getConnection().prepareStatement("INSERT INTO menu(name,description,price,type) VALUES (?,?,?,?)");
-                                con.setString(1, map.get("name").getText());
-                                con.setString(2, map.get("description").getText());
-                                con.setString(3, map.get("price").getText());
-                                con.setString(4, map.get("type").getText());
-                                con.executeUpdate();
-                                tablePane.reprintMe();
+                            switch (tablename) {
+                                case "impiegati":
+                                    con = mgr.getConnection().prepareStatement("USE ristorante");
+                                    con.executeQuery();
+                                    con = mgr.getConnection().prepareStatement("INSERT INTO impiegati(username,password,ruolo) VALUES (?,?,?)");
+                                    con.setString(1, map.get("username").getText());
+                                    con.setString(2, map.get("password").getText());
+                                    con.setString(3, map.get("ruolo").getText());
+                                    con.executeUpdate();
+                                    break;
+                                case "menu":
+                                    con = mgr.getConnection().prepareStatement("USE ristorante");
+                                    con.executeQuery();
+                                    con = mgr.getConnection().prepareStatement("INSERT INTO menu(name,description,price,type) VALUES (?,?,?,?)");
+                                    con.setString(1, map.get("name").getText());
+                                    con.setString(2, map.get("description").getText());
+                                    con.setString(3, map.get("price").getText());
+                                    con.setString(4, map.get("type").getText());
+                                    con.executeUpdate();
+                                    break;
                             }
-                        tablePane.reprintMe();
-                        } catch (SQLException ex) {
+                        } catch (SQLException ex){
                             Logger.getLogger(FunctionalPanel.class.getName()).log(Level.SEVERE, null, ex);
                         }
                     }
@@ -97,25 +100,72 @@ public class buttonLayout extends JPanel{
         
         b3.addActionListener(new ActionListener(){
             @Override
-            public void actionPerformed(ActionEvent e) {
-                String colName = tablePane.getSelectedColumnName();     //per cancellare una entrata si deve selezionare la colonna id
-                String str = map.get(colName).getText();
-                int id = Integer.parseInt(str);
-                               
+            public void actionPerformed(ActionEvent e){
+                int id=0;
+                String str = map.get("id").getText();
+                id = Integer.parseInt(str);
                 if(!str.equals("")){
                     try {
                         PreparedStatement con = mgr.getConnection().prepareStatement("USE ristorante");
                         con.executeQuery();
-                        con = mgr.getConnection().prepareStatement("DELETE FROM impiegati WHERE id=?");
-                        //con.setString(1, tablename);
-                        //con.setString(2, colName);
-                        con.setInt(1, id);
-                        con.executeUpdate();
-                    } catch (SQLException ex) {
+                        if(!checkExistingId(mgr.getConnection(),tablename,id)){
+                            con = mgr.getConnection().prepareStatement("DELETE FROM impiegati WHERE id=?");
+                            con.setInt(1, id);
+                            con.executeUpdate();
+                        }else{
+                            JOptionPane.showMessageDialog(frame,"ID inesistente","Errore",JOptionPane.ERROR_MESSAGE);
+                        }
+                    } catch (SQLException ex){
                         Logger.getLogger(FunctionalPanel.class.getName()).log(Level.SEVERE, null, ex);
                     }
+                }else{
+                    JOptionPane.showMessageDialog(frame,"Specificare un id da cancellare","Errore",JOptionPane.ERROR_MESSAGE);
                 }
             }
         });
     }
+    
+    private boolean checkExistingId(Connection con,String nametable,int id){
+        boolean flag=false;
+        ResultSet rs = null;
+        switch(nametable){
+            case "impiegati":
+                try {
+                    PreparedStatement stat=con.prepareStatement("SELECT * FROM impiegati WHERE id=?");
+                    //stat.setString(1, nametable);
+                    stat.setInt(1, id);
+                    rs=stat.executeQuery();
+                } catch (SQLException ex) {
+                    Logger.getLogger(buttonLayout.class.getName()).log(Level.SEVERE, null, ex);
+                }
+                break;
+            case "menu":
+                try {
+                    PreparedStatement stat=con.prepareStatement("SELECT * FROM menu WHERE id=?");
+                    //stat.setString(1, nametable);
+                    stat.setInt(1, id);
+                    rs=stat.executeQuery();
+                } catch (SQLException ex) {
+                    Logger.getLogger(buttonLayout.class.getName()).log(Level.SEVERE, null, ex);
+                }
+                break;
+        }
+        return isResultSetEmpty(rs);
+    }
+    
+    //false=resultset non vuoto
+    private boolean isResultSetEmpty(ResultSet rs){
+        boolean flag=false;
+        try {
+            if(rs.next()){
+                flag=false;
+            }else{
+                flag=true;
+            }
+        } catch (SQLException ex) {
+            Logger.getLogger(buttonLayout.class.getName()).log(Level.SEVERE, null, ex);
+        }
+        return flag;
+    }
+    
 }
